Add TextFitArray
Add a new function to satisfy the UX that
operates TextFit by considering the PointSize and MinLineSize of the text.

For example, TextFit should be able to find the one that
fits among [PointSize 24 + MinLineSize 40] or [PointSize 28 + MinLineSize 44].

But the previous TextFit only considers PointSize.

TextFitArray can solve this problem,
and this implemented it to respond to additional requirements without modifying the API.

https://review.tizen.org/gerrit/#/c/platform/core/uifw/dali-toolkit/+/299557/
https://review.tizen.org/gerrit/#/c/platform/core/uifw/dali-csharp-binder/+/299690/

Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>