Fix defect that is detected by static analisys tool
There is a portential problem with the code involving the variable 'e.OldItems'.
This variable is the result of a method invocation, and the tool is warning that the method
might return a null value, which can cause deferencing a null reference.
It might cause to crash or behave unpredictably.

Therefore, to prevent this, the variable 'e.OldItems' should be checked if it is null or not before use.