[NUI] Apply CornerRadius to WidgetView
The CornerRadius function previously supported in View has been modified to also apply to WidgetView.
For this behavior, WidgetView has an internal visual and its structure is changed.
In this case,  WidgetView no longer generates shaders and rendering directly.

I've added a sample for testing the corner radius.
You can use it as follows:
widgetView.CornerRadius = 22;