[NUI] Add ValuePolicy for BindableProperty
"ALL" default BindableProperty.PropertyChanged events what NUI using, dont use 'oldValue'.

But due to that usecase, we always call DefaultValueCreator when we try to set value.

To make NUI platform faster, let we ignore that property getter when we only use setter.

Do not call default creator when we set value

Currently, we only apply this feature to View's bindable property.
But in future, we need to create all bindable property policy as IgnoreOldValueWhenSetValue.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>

[NUI] Set ValuePolicy for ImageView and TextLabel

Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>

[NUI] Minor fix to IgnoreValueGetter() method

Let we use switch case, to prepare additional policy.

Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>

[NUI] Apply ValuePolicy as IgnoreOldValueWhenSetValue for all BaseComponents

Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>

[NUI] Apply ValuePolicy as IgnoreOldValueWhenSetValue for all non-BaseComponents

Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>

[NUI] Apply some BindableProperties who using delegater instead of lambda

Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>

[NUI] Add useful policy + Apply comments

Let we make new policy that will not ask to default creator anymore
if we set before, or get from native before.

It will also optimize the GetValue cases in future.

Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>