[NUI][API11] Cherry-pick potential issue at DisposeQueue
This is a combination of 2 commits.

- This is the 1st commit message:

[NUI] Resolve the potential issue when DisposeQueued handle get

It is possible that DisposeQueued handle be used as input of GetManagedBaseHandleFromRefObject.

In this case, even if app hold that BaseHandle, it will be Disposed by DisposeQueue.
If then the returned item become disposed, and it might breakdown app logics.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>

- This is the commit message 2:

[NUI] Unregister disposed / disposequeued BaseHandle and register again if need

It is possible that dispose queued basehandle try to get / register again
due to some reason. Like, Native side object is alived but C# side has no reference.

We already to to detect that cases. But if BaseHandle added at DisposeQueue,
we cannot avoid to call the Dispose(); after some event loops later, and then
Unregister the handle.

To avoid some error cases, let we Unregister the basehandle synchronously
and then re-register base on the new BaseHandle.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>