[NUI] Implement VisualObject feature for View using Visual feature
It was hard to use Visual feature for `BaseComponents.View`

We only allow to use `VisualBase` only for `VisualView`, with name.
But there was some user side hardness when they want to use it

1. Subclass of View cannot use Visual feature
2. Some ImageVisual feature doesn't allow to used another visuals.

To resolve this issue,
let we make new class, named `Visuals.VisualBase` and let View use it.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>