[NUI] Unregister disposed / disposequeued BaseHandle and register again if need
It is possible that dispose queued basehandle try to get / register again
due to some reason. Like, Native side object is alived but C# side has no reference.

We already to to detect that cases. But if BaseHandle added at DisposeQueue,
we cannot avoid to call the Dispose(); after some event loops later, and then
Unregister the handle.

To avoid some error cases, let we Unregister the basehandle synchronously
and then re-register base on the new BaseHandle.

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>