[NUI] Support Shadow and ColorVisual cutout policy
Let we support to BoxShadow and ColorVisual, to ignore rendering inside of
View area.

None is default, same as previous logics.
CutoutView is cutout render area as bounding box of view. It is faster than CutoutViewWithCornerRadius.
CutoutWithCornerRadius is cutout render area consider the corner radius of view.

Required dali patch :
https://review.tizen.org/gerrit/c/platform/core/uifw/dali-toolkit/+/314255

Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>