[NUI] Fix SVACE issues
1) DF240922-00003
https://sa.sec.samsung.net/dm/tizen90/sb2/main/review#PRJID=1&WGID=43585
false alarm
At line 755, a null check has been already performed ; the item "e.Proposed" which was pointed out as an problem cannot become null.
As such, this issue does not require correction as it is a false alarm.

2) DF240907-00723
https://sa.sec.samsung.net/dm/tizen90/sb2/main/review#PRJID=1&WGID=176282
fixed

3) DF240907-00695
https://sa.sec.samsung.net/dm/tizen90/sb2/main/review#PRJID=1&WGID=176248
fixed

4) DF240906-00277
https://sa.sec.samsung.net/dm/tizen90/sb2/main/review#PRJID=1&WGID=174470
fixed

5) DF240905-01180
https://sa.sec.samsung.net/dm/tizen90/sb2/main/review#PRJID=1&WGID=43601

At line 1014, a null check has been already performed; if it is null, an exception will be thrown.
Therefore, at line 1018, the item which was pointed out as an problem cannot become null.
As such, this issue does not require correction as it is a false alarm.

6) DF240821-00053
https://sa.sec.samsung.net/dm/tizen90/sb2/main/review#PRJID=1&WGID=163282
fixed

7) DF240716-01912
https://sa.sec.samsung.net/dm/tizen90/sb2/main/review#PRJID=1&WGID=163084
already fixed in the previous patch