[NUI] Introduce new binding style.
The binding in NUI cause performance and memory issue,
so we plan to deprecate them by setting IsUsingXaml disable.

This patch is aim to introduce new style of binding,
which do not causing performance and memory seriously,
and user can add binding property on existing view class,
so that we do not need to add every property to bindable.