[NUI] Add StartOffset of GradientVisual
### Description of Change ###
Add StartOffset of GradientVisual.
Start offset is the offset value that shifts the starting position of the gradient.
0.0 is start position of gradient, 1.0 is end position of gradient.
It is possible to set the startOffset outside the [0, 1] range.
For example, you can animate it from 0.5 to 1.5.

How to use:
Animation anim = new Animation(1000);
anim.AnimateTo(bgView, "gradient.StartOffset", 1.0f);
anim.Play();