[NUI] Add GetInsets to Window
Window insets are the inner size between window content and window edge.

Window insets are required when system UI such as status bar, virtual
keyboard, or clipboard appears on window to avoid window content
obscured by them.

To provide the combined window insets for the specified parts of the
system UI, GetInsets is added to Window.

The parameter insetsFlags of GetInsets is a bitwise combination of
InsetsPartFlags values specifying which window insets parts to include.