Apply of Lazy Loading to ULocale Class
In the constructor of ULocale, there is a logic
to import member variables using api.

Since only some member variables are used in the existing code,
I think it is more efficient to apply lazy loading so that
member variable can be imported only once when necessary
than to call them using api at the time of creation.

Signed-off-by: Hyeon-Uk <rlagusdnr120@gmail.com>