[NUI.Gadget] Add new methods for adding delay
The NUIGadgetManager.ResumeDelayed() and the NUIGadgetManager.PauseDelayed()
are added to add the delay. When Application repeatedly calls Resume() and
Pause(), Gadget's OnResume() and OnPause() are called repeatedly. In this case,
if key event is passed, it will not be processed immediately.
All event of glib must be processed to operate in a structure in which
fd handler of ecore is processed.
This patch is to solve the problem.

Signed-off-by: Hwankyu Jhun <h.jhun@samsung.com>