[NUI] Fix GridLayout's infinite layout calculation for stretch fill
Previously, GridLayout calculated layout infinitely if it had children
which had stretch fill option.

Because children with stretch fill option are measured again with
new LayoutWidth/Height to calculate the grand children's size based on
the filled size.

When children's LayoutWidth/Height are changed, RequestLayout is called
for the children and RequestLayout is propagated to GridLayout.

This causes that GridLayout's Measure and Layout are called infinitely.

Now, GridLayout sets LayoutRequested true not to propagate RequestLayout
from children and restores LayoutRequested.