Make GpioPin constructor Dispose() only on errors
The intention of the code seems to be to close the handle on errors,
since there is some initialization in the constructor which might need
to be undoed. However, finally is not a correct way to do so, since it
will also execute the code on error. Instead, we make the disposal
happen in an exception handler and make it rethrow the exception.