Fix the starvation of lifecycle event
If the lifecycle event throws an exception, the ProcessLifecycleEvent function
is terminated without changing the Critical Section Flag(__processing)

Since then, we can not run any other lifecycle events that come in.

So, even if exception is thrown, flag can be changed through finally,
and lock is used rather than flag value for concurrency.

Signed-off-by: Hyeon-Uk <rlagusdnr120@gmail.com>